requirements.txt
openai>=1.0.0
Pillow==10.4.0
requests==2.31.0
pyyaml==6.0.2
python-dotenv==1.0.1
typer==0.9.0
rich==13.3.0
click>=8.0.0,<8.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
psutil>=5.9.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.9.0
textblob>=0.17.0
joblib>=1.1.0

# Enterprise Task 3 Dependencies
aiohttp>=3.12.0
prometheus-client>=0.22.0
watchdog>=3.0.0
dataclasses-json>=0.6.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
bandit>=1.7.0
safety>=2.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0